% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelyr.R
\name{day_name_of_week}
\alias{day_name_of_week}
\title{day_name_of_week}
\usage{
day_name_of_week(userDate)
}
\arguments{
\item{userDate}{in YYYY/MM/DD format}
}
\value{
One of the following { c("Sunday","Monday","Tuesday","Wednesday","Thursday","Friday","Saturday")}
}
\description{
Get day name of week
}
\details{
Given a date return day name of week
}
\examples{
userDate <- "2000/01/01"
day_name_of_week(userDate)
}
\author{
armen donigian
}

