% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelyr.R
\name{day_of_week}
\alias{day_of_week}
\title{day_of_week}
\usage{
day_of_week(userDate)
}
\arguments{
\item{userDate}{in YYYY/MM/DD format}
}
\value{
One of the following { c("1","2","3","4","5","6","7")}
}
\description{
Get day of week
}
\details{
Given a date return day of week
}
\examples{
userDate <- "2000/01/01"
day_of_week(userDate)
}
\author{
armen donigian
}

