% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelyr.R
\name{day_of_year}
\alias{day_of_year}
\title{day_of_year}
\usage{
day_of_year(userDate)
}
\arguments{
\item{userDate}{in YYYY/MM/DD format}
}
\value{
One of the following { c(1:365)}
}
\description{
Get day of year
}
\details{
Given a date return day of year
}
\examples{
userDate <- "2000/01/01"
day_of_year(userDate)
}
\author{
armen donigian
}

