% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelyr.R
\name{fiscal_quarter}
\alias{fiscal_quarter}
\title{fiscal_quarter}
\usage{
fiscal_quarter(userDate)
}
\arguments{
\item{userDate}{in YYYY/MM/DD format}
}
\value{
One of the following { c("1","2","3","4","5","6","7")}
}
\description{
Get fiscal month of year
}
\details{
Given a date return fiscal month of year
}
\examples{
userDate <- "2000/01/01"
fiscal_quarter(userDate)
}
\author{
armen donigian
}

