% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelyr.R
\name{is.last_day_of_month}
\alias{is.last_day_of_month}
\title{is.last_day_of_month}
\usage{
is.last_day_of_month(userDate)
}
\arguments{
\item{userDate}{in YYYY/MM/DD format}
}
\value{
TRUE or FALSE
}
\description{
Is a given date last day of month
}
\details{
Given a date return TRUE if date is last day of month, FALSE otherwise
}
\examples{
userDate <- "2000/01/01"
is.last_day_of_month(userDate)
}
\author{
armen donigian
}

