% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelyr.R
\name{is.weekend}
\alias{is.weekend}
\title{is.weekend}
\usage{
is.weekend(userDate)
}
\arguments{
\item{userDate}{in YYYY/MM/DD format}
}
\value{
TRUE or FALSE
}
\description{
Is a given date a weekend
}
\details{
Given a date returns TRUE if Weekend, FALSE otherwise
}
\examples{
userDate <- "2000/01/01"
is.weekend(userDate)
}
\author{
armen donigian
}

